requirements.txt
SQLAlchemy==2.0.23
PyMySQL==1.1.0
python-dotenv==1.0.0
tabulate==0.9.0
datetime
